+++ pkg/utils/flow_utils.go
@@ -256,14 +256,14 @@ func DNSRcodeToString(f *flow.Flow) string {
 }
 
 // AddPacketSize adds the packet size to the flow's metadata.
-func AddPacketSize(meta *RetinaMetadata, packetSize uint64) {
+func AddPacketSize(meta *RetinaMetadata, packetSize uint32) {
 	if meta == nil {
 		return
 	}
 	meta.Bytes = packetSize
 }
 
-func PacketSize(f *flow.Flow) uint64 {
+func PacketSize(f *flow.Flow) uint32 {
 	if f.Extensions == nil {
 		return 0
 	}
@@ -273,7 +273,7 @@ func PacketSize(f *flow.Flow) uint64 {
 }
 
 // AddDropReason adds the drop reason to the flow's metadata.
-func AddDropReason(f *flow.Flow, meta *RetinaMetadata, dropReason uint32) {
+func AddDropReason(f *flow.Flow, meta *RetinaMetadata, dropReason uint16) {
 	if f == nil || meta == nil {
 		return
 	}
